Prism/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
supabase>=2.7.0
httpx>=0.27.0
pydantic>=2.7.0
torch>=2.0.0
transformers>=4.40.0
selenium>=4.20.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
